Deployment/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
tensorflow==2.19.0
Pillow==10.0.0
numpy==2.1.3
gunicorn==21.2.0
h5py==3.11.0
Flask-Cors==4.0.0